global-pulse24/server/app.py
from flask import Flask, request, jsonify
from flask_pymongo import PyMongo
from flask_cors import CORS
from bson.objectid import ObjectId
import os
from datetime import datetime
from dotenv import load_dotenv

# Load Environment Variables
load_dotenv()

app = Flask(__name__)

# Configuration
app.config["MONGO_URI"] = os.getenv("MONGO_URI", "mongodb://localhost:27017/globalpulse24")
app.config["SECRET_KEY"] = os.getenv("SECRET_KEY", "super_secret_dev_key")
ADMIN_USER = os.getenv("ADMIN_USER", "admin")
ADMIN_PASS = os.getenv("ADMIN_PASS", "password")

# Simple In-Memory Token Storage (restarts with server)
current_valid_token = None

# Extensions
mongo = PyMongo(app)
CORS(app) # Allow all for now, can restrict to specific domains in production

# --- Routes ---

@app.route('/health', methods=['GET'])
def health_check():
    return jsonify({"status": "healthy", "service": "GlobalPulse24 Backend"}), 200

# 1. Login Route
@app.route('/login', methods=['POST'])
def login():
    data = request.json
    username = data.get('username')
    password = data.get('password')
    # Server-Side verification of the client's flag (as per prompt instructions)
    captcha_verified = data.get('captcha_verified', False)

    if not captcha_verified:
         return jsonify({"success": False, "message": "CAPTCHA verification failed"}), 400

    if username == ADMIN_USER and password == ADMIN_PASS:
        # Generate a random session token
        # In production this would be signed (JWT), but UUID is better than static
        import uuid
        token = str(uuid.uuid4())
        
        # NOTE: Ideally store this token in DB/Redis to validate later. 
        # For this student project, we will use a simplified approach where
        # valid tokens are just logged or we trust the client to have "a" token 
        # (Wait, prompt says "Protected Route: Check for this session token").
        # To strictly follow "Check for THIS session token", we need to store it.
        # But to keep it simple as requested ("Student Disclaimer"), we might valid check purely logic 
        # or implement a simple valid list.
        # Let's keep it robust enough: matching prompt "secure, short-lived session token".
        # We'll just define a valid session list or valid current token in memory (simple server).
        global current_valid_token 
        current_valid_token = token
        
        return jsonify({"success": True, "message": "Login successful", "token": token}), 200
    else:
        return jsonify({"success": False, "message": "Invalid credentials"}), 401

# 2. Public News Route (GET)
@app.route('/api/news', methods=['GET'])
def get_news():
    try:
        # Sort by createdAt descending (-1)
        news_cursor = mongo.db.news.find().sort("createdAt", -1)
        news_list = []
        for doc in news_cursor:
            news_list.append({
                "_id": str(doc["_id"]),
                "title": doc.get("title"),
                "content": doc.get("content"),
                "imageUrl": doc.get("imageUrl"),
                "source": doc.get("source"),
                "sourceUrl": doc.get("sourceUrl"),
                "category": doc.get("category", "World"),
                "createdAt": doc.get("createdAt").isoformat() if doc.get("createdAt") else None
            })
        return jsonify(news_list), 200
    except Exception as e:
        return jsonify({"error": str(e)}), 500

# 3. Secure Add News Route (POST) - Prompt requested '/add-news'
@app.route('/add-news', methods=['POST'])
def add_news():
    # Simple Auth Check (Header)
    auth_header = request.headers.get('Authorization')
    token = auth_header.split(" ")[1] if auth_header and len(auth_header.split(" ")) > 1 else None
    
    if not token or token != current_valid_token:
        return jsonify({"error": "Unauthorized"}), 401

    try:
        data = request.json
        new_article = {
            "title": data.get("title"),
            "content": data.get("content"),
            "imageUrl": data.get("imageUrl"),
            "source": data.get("source"),
            "sourceUrl": data.get("sourceUrl"),
            "category": data.get("category", "World"),
            "createdAt": datetime.utcnow()
        }
        
        result = mongo.db.news.insert_one(new_article)
        new_article["_id"] = str(result.inserted_id)
        
        return jsonify({"success": True, "article": new_article}), 201
    except Exception as e:
        return jsonify({"error": str(e)}), 400

# 4. Alias for compatibility (POST /api/news)
@app.route('/api/news', methods=['POST'])
def create_news_alias():
    return add_news()

if __name__ == '__main__':
    port = int(os.environ.get("PORT", 5000))
    app.run(host='0.0.0.0', port=port, debug=True)
