global-pulse24/server/requirements.txt
flask
flask-cors
flask-pymongo
python-dotenv
pymongo
